// src/lib/server/graphqlServer/types/generated-resolvers.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  AccountNumber: any;
  BigInt: any;
  Byte: any;
  CountryCode: any;
  Currency: any;
  DID: any;
  Date: any;
  DateTime: any;
  Duration: any;
  EmailAddress: any;
  GUID: any;
  HSL: any;
  HSLA: any;
  HexColorCode: any;
  Hexadecimal: any;
  IBAN: any;
  IPv4: any;
  IPv6: any;
  ISBN: any;
  ISO8601Duration: any;
  JSON: any;
  JSONObject: any;
  JWT: any;
  Latitude: any;
  LocalDate: any;
  LocalEndTime: any;
  LocalTime: any;
  Locale: any;
  Long: any;
  Longitude: any;
  MAC: any;
  NegativeFloat: any;
  NegativeInt: any;
  NonEmptyString: any;
  NonNegativeFloat: any;
  NonNegativeInt: any;
  NonPositiveFloat: any;
  NonPositiveInt: any;
  ObjectID: any;
  PhoneNumber: any;
  Port: any;
  PositiveFloat: any;
  PositiveInt: any;
  PostalCode: any;
  RGB: any;
  RGBA: any;
  RoutingNumber: any;
  SafeInt: any;
  Time: any;
  TimeZone: any;
  Timestamp: any;
  URL: any;
  USCurrency: any;
  UUID: any;
  UnsignedFloat: any;
  UnsignedInt: any;
  UtcOffset: any;
  Void: any;
};
// ...
export type AccountGroupingFilter = {
  active?: InputMaybe<BooleanFilter>;
  deleted?: InputMaybe<BooleanFilter>;
  disabled?: InputMaybe<BooleanFilter>;
  title?: InputMaybe<StringFilter>;
};

export type AccountGroupingIdFilter = {
  equals?: InputMaybe<Scalars['UUID']>;
  in?: InputMaybe<Array<InputMaybe<Scalars['UUID']>>>;
  not?: InputMaybe<Scalars['UUID']>;
  notIn?: InputMaybe<Array<InputMaybe<Scalars['UUID']>>>;
};
// ...
export type BooleanFilter = {
  equals?: InputMaybe<Scalars['Boolean']>;
  not?: InputMaybe<Scalars['Boolean']>;
};
// ...
export type BudgetFilter = {
  accountGrouping?: InputMaybe<AccountGroupingFilter>;
  accountGroupingId?: InputMaybe<AccountGroupingIdFilter>;
  active?: InputMaybe<BooleanFilter>;
  allowUpdate?: InputMaybe<BooleanFilter>;
  deleted?: InputMaybe<BooleanFilter>;
  disabled?: InputMaybe<BooleanFilter>;
  id?: InputMaybe<StringFilter>;
  status?: InputMaybe<StatusFilter>;
  title?: InputMaybe<StringFilter>;
};
// ...
export type StatusEnum =
  | 'Active'
  | 'Deleted'
  | 'Disabled';

export type StatusFilter = {
  equals?: InputMaybe<StatusEnum>;
  in?: InputMaybe<Array<InputMaybe<StatusEnum>>>;
  not?: InputMaybe<StatusEnum>;
  notIn?: InputMaybe<Array<InputMaybe<StatusEnum>>>;
};

export type StringFilter = {
  contains?: InputMaybe<Scalars['String']>;
  endsWith?: InputMaybe<Scalars['String']>;
  equals?: InputMaybe<Scalars['String']>;
  in?: InputMaybe<Array<InputMaybe<Scalars['String']>>>;
  not?: InputMaybe<Scalars['String']>;
  notIn?: InputMaybe<Array<InputMaybe<Scalars['String']>>>;
  startsWith?: InputMaybe<Scalars['String']>;
};